Fix uint32_t integer overflow in TransposeDecoder state machine resize
In ParseStateMachine(), `state_machine_size` is a `uint32_t` read from
attacker-controlled varint-encoded input. The expression:

  context.state_machine_nodes.resize(state_machine_size + 0xff);

performs the addition in 32-bit unsigned arithmetic. When
`state_machine_size >= 0xFFFFFF01`, the sum wraps around to a value
near zero (e.g. 0xFFFFFF01 + 0xFF = 0x100000000 truncates to 0x0),
causing `resize(0)`. The three subsequent loops then iterate
`state_machine_size` times writing into the empty vector, producing
a heap buffer overflow (CWE-122 via CWE-190).

Fix: widen `state_machine_size` to `size_t` before the addition so the
arithmetic is performed in 64-bit and cannot wrap. Remove the
`// TODO: Handle overflow.` note that acknowledged the issue.

Resolves OSS VRP issue 495353866.

+++ riegeli/chunk_encoding/transpose_decoder.cc
@@ -555,8 +555,10 @@ inline bool TransposeDecoder::Parse(Context& context, Reader& src,
         absl::InvalidArgumentError("Reading state machine size failed")));
   }
   // Additional `0xff` nodes to correctly handle invalid/malicious inputs.
-  // TODO: Handle overflow.
-  context.state_machine_nodes.resize(state_machine_size + 0xff);
+  // Cast to `size_t` before addition to prevent `uint32_t` wraparound when
+  // `state_machine_size` is near UINT32_MAX.
+  context.state_machine_nodes.resize(static_cast<size_t>(state_machine_size) +
+                                     0xff);
   if (projection_enabled) context.node_templates.resize(state_machine_size);
   std::vector<StateMachineNode>& state_machine_nodes =
       context.state_machine_nodes;
